grpc client and server launch method adjustments

+++ drivers/shared/executor/executor.go
@@ -166,7 +166,7 @@ type ExecCommand struct {
 
 	// OOMScoreAdj allows setting oom_score_adj (likelihood of process being
 	// OOM killed) on Linux systems
-	OOMScoreAdj int
+	OOMScoreAdj int64
 }
 
 func (c *ExecCommand) getCgroupOr(controller, fallback string) string {
